requirements.txt
clams-python==1.0.2
scenedetect[opencv-headless]==0.6.1
